+++ ChannelIO/Source/Utils/Message Format/MarkdownParser.swift
@@ -38,7 +38,7 @@ open class MarkdownParser {
   /// Enables or disables detection of URLs even without Markdown format
   open var automaticLinkDetectionEnabled: Bool = true
   open let font: UIFont
-  let emojiFont: UIFont = UIFont.systemFont(ofSize: 36)
+  let emojiFont: UIFont = UIFont.systemFont(ofSize: 40)
   
   // MARK: Initializer
   public init(font: UIFont = UIFont.systemFont(ofSize: UIFont.smallSystemFontSize),